Got a basic setup where I have various pages querying notes. Each page allows for adding new notes. Each page also allows for editing and deleting notes through context menu actions.

+++ spaces/common/CommonSpace.ts
@@ -78,20 +78,22 @@ export class CommonSpace implements LogicalSpace {
 
 
     buildNoteActionsMenu(note: Note, menuBuilder: NoteActionsMenuBuilder, notu: Notu) {
-        menuBuilder.addToTopOfStart(
-            new NoteAction('Edit',
-                () => Promise.resolve(new ShowEditorAction(note))
-            )
-        );
-        menuBuilder.addToBottomOfEnd(
-            new NoteAction('Delete',
-                async () => {
-                    await notu.saveNotes([note.delete()]);
-                    return new RefreshAction();
-                },
-                true
-            )
-        );
+        if (note.ownTag?.isInternal != true) {
+            menuBuilder.addToTopOfStart(
+                new NoteAction('Edit',
+                    () => Promise.resolve(new ShowEditorAction(note))
+                )
+            );
+            menuBuilder.addToBottomOfEnd(
+                new NoteAction('Delete',
+                    async () => {
+                        await notu.saveNotes([note.delete()]);
+                        return new RefreshAction();
+                    },
+                    true
+                )
+            );
+        }
     }
